split `_internal` into two separate files so only the `noUncheckedIndexedAccess` checker needs to be a `.d.ts` file and prettier-ignore it due to syntax error
