+++ .prettierignore
@@ -2,3 +2,4 @@ node_modules
 dist
 package.json
 package-lock.json
+src/types/_internal/noUncheckedIndexedAccessChecker.d.ts
